+++ src/components/Widgets/FeedbackWidget/FeedbackCard.js
@@ -2,13 +2,9 @@ import React from 'react';
 import LeafygreenCard from '@leafygreen-ui/card';
 import styled from '@emotion/styled';
 
-import CloseButton from './components/CloseButton';
 import ProgressBar from './components/ProgressBar';
-import { useFeedbackState } from './context';
 
 export default function FeedbackCard({ isOpen, children }) {
-  const { abandon } = useFeedbackState();
-
   return (
     isOpen && (
       <Floating>
@@ -30,12 +26,8 @@ const Floating = styled.div`
 const Card = styled(LeafygreenCard)`
   width: 230px;
 `;
-const CardHeader = styled.div`
-  display: flex;
-  justify-content: flex-end;
-  margin-top: 10px;
-  margin-right: 10px;
-`;
+
 const Content = styled.div`
-  margin: 0px 24px 35px 24px;
+  margin: 0px 24px 35px;
 `;
+
